third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-uncaught-promise-on-pause-expected.txt

CONSOLE ERROR: line 16: Uncaught (in promise) Error: inspector.err1
CONSOLE ERROR: line 34: Uncaught (in promise) Error: inspector.err2
Tests uncaught promise rejections fired during pause.

Set timer for test function.
debugger-uncaught-promise-on-pause.html:10 Console was cleared

    debugger-uncaught-promise-on-pause.html:16 Uncaught (in promise) Error: inspector.err1 {stack: (...), message: "inspector.err1"}
        message: "inspector.err1"
        stack: (...)
        get stack: function () { [native code] }
        set stack: function () { [native code] }
        __proto__: d
        debugger-uncaught-promise-on-pause.html:16 runPromises
        debugger-uncaught-promise-on-pause.html:41 timeout

    debugger-uncaught-promise-on-pause.html:34 Uncaught (in promise) Error: inspector.err2 {stack: (...), message: "inspector.err2"}
        message: "inspector.err2"
        stack: (...)
        get stack: function () { [native code] }
        set stack: function () { [native code] }
        __proto__: d
        debugger-uncaught-promise-on-pause.html:34 runPromises
        debugger-uncaught-promise-on-pause.html:41 timeout
